// lib/api/schema.ts
export interface components {
    schemas: {
// ...
        ProviderPublic: {
            /** Slug */
            slug: string;
            /** Biz Name */
            biz_name: string;
            /**
             * Verification Status
             * @default pending
             */
            verification_status: string;
            /**
             * Is Open
             * @default true
             */
            is_open: boolean;
            /**
             * Is Paused
             * @default false
             */
            is_paused: boolean;
            /** Join Radius M */
            join_radius_m?: number | null;
            /** Latitude */
            latitude?: number | null;
            /** Longitude */
            longitude?: number | null;
            /**
             * Is Private
             * @default false
             */
            is_private: boolean;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Created At */
            created_at?: string | null;
            /**
             * Ratings Count
             * @default 0
             */
            ratings_count: number;
            /** Rating Avg */
            rating_avg?: number | null;
        };
        /** ProviderRatingSummary */
        ProviderRatingSummary: {
            /**
             * Provider Id
             * Format: uuid
             */
            provider_id: string;
            /**
             * Ratings Count
             * @default 0
             */
            ratings_count: number;
            /** Rating Avg */
            rating_avg?: number | null;
            /** Estimate Accuracy Rate */
            estimate_accuracy_rate?: number | null;
        };
// ...
        ServiceItemPublic: {
            /** Name */
            name: string;
            /** Avg Duration Minutes */
            avg_duration_minutes: number;
            /** Price */
            price: string;
            /**
             * Is Active
             * @default true
             */
            is_active: boolean;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /**
             * Provider Id
             * Format: uuid
             */
            provider_id: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}